api-clientes/requirements.txt
fastapi[all]
SQLAlchemy
psycopg2-binary
pydantic
uvicorn
python-dotenv
pymysql
cryptography
# mysqlclient
# asyncmy
# greenlet